+++ rao-runner-api/src/test/resources/timeCoupledRaoResponseMessage.json
@@ -3,7 +3,6 @@
     "type": "time-coupled-rao-response",
     "id": "id",
     "attributes": {
-      "instant": "instant",
       "networksWithPraFileUrl": "networksWithPraFileUrl",
       "raoResultsFileUrl": "raoResultsFileUrl",
       "computationStartInstant": 1637052884.944727,
