--- rao-runner-api/src/test/resources/timeCoupledRaoResponseMessageNullInstant.json
@@ -1,11 +1,0 @@
-{
-  "data": {
-    "type": "time-coupled-rao-response",
-    "id": "id",
-    "attributes": {
-      "instant": null,
-      "networksWithPraFileUrl": "networksWithPraFileUrl",
-      "raoResultsFileUrl": "raoResultsFileUrl"
-    }
-  }
-}
