+++ rao-runner-spring-boot-starter/src/test/resources/timeCoupledRaoRequestMessage.json
@@ -3,7 +3,6 @@
     "type": "time-coupled-rao-request",
     "id": "id",
     "attributes": {
-      "instant": "instant",
       "icsFileUrl": "icsFileUrl",
       "timedInputs": [{
         "timestamp": "2025-12-31T12:34Z",
