+++ rao-runner-spring-boot-starter/src/test/resources/timeCoupledRaoResponseMessage.json
@@ -3,7 +3,6 @@
     "type": "time-coupled-rao-response",
     "id": "id",
     "attributes": {
-      "instant": "instant",
       "networksWithPraFileUrl": "networksWithPraFileUrl",
       "raoResultsFileUrl": "raoResultsFileUrl"
     }
